+++ webapp/src/main/resources/static/js/NavBar.js
@@ -1,4 +1,5 @@
 import React, {Component} from "react";
+import {Link} from "react-router-dom";
 
 class NavBar extends Component {
 
@@ -12,6 +13,13 @@ class NavBar extends Component {
                             <span className="sr-only">Toggle navigation</span>
                         </button>
                         <a className="navbar-brand" href="#">Yet Another Issue Tracker</a>
+                    </div>
+                    <div className="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+                        <ul className="nav navbar-nav">
+                            <li>
+                                <Link to={'/login'}>Login</Link>
+                            </li>
+                        </ul>
                     </div>
                 </div>
             </nav>
